packages/agents/requirements.txt
firebase-admin
livekit-agents[silero,turn-detector]
python-dotenv